requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
chromadb==0.4.18
pandas==2.0.3
numpy==1.24.3
streamlit==1.28.0
scikit-learn==1.3.2
openai==1.56.1
openpyxl==3.1.2
groq==0.33.0
sentence-transformers==2.2.2
torch==2.0.1
google-generativeai>=0.8.0
